actors/python-sound/requirements.txt
pyaudio
paho-mqtt